Add support for mining the "center" source keeper room surrounded by source keepers, but without any directly guarding sources in that room, and implement urgent tower code fix!
Tower code fix kind of ended up in here by accident, but eh, it'll be fine.

Changes:
- Add an INVADER_USERNAME constant instead of using "Invader" everywhere
- Have invaders in source keeper rooms counted as hostiles, while still ignoring other player's non-military creeps and source keepers
- Add a "SK_LAIR_SOURCE_NOTED" flag which causes pathfinding to completely avoid a 4-range radius from the flag, even in rooms we don't have vision in
- Split "full_miner" creep base into "3000miner" and "4500miner", with 5 and 8 work parts respectively
- Spawn more haulers for remote mining flags with a "sk_room" memory value set to a truthy value
- Spawn a 4500miner for remote mining flags with a "sk_room" memory value set to a truthy value
- TODO: auto-determine if a room is an SK room based off of room_x % 10 and room_y % 10
- Only spawn reservers for mining flags without a "sk_room" memory value
- Fix up creep passingby-road-building to target the most built roads (and lowest hit roads for repairing) first, so as to quickly build some roads for faster movement, and also to free up more construction sites for future roads more quickly
- Fix tower code by actually specifying an argument to tower.attack()

+++ src/roles/military.py
@@ -4,7 +4,7 @@ import random
 
 import autoactions
 import flags
-from constants import target_single_flag, role_td_healer
+from constants import target_single_flag, role_td_healer, INVADER_USERNAME
 from control import pathdef
 from role_base import RoleBase
 from tools import profiling
@@ -61,7 +61,7 @@ class RoleDefender(RoleBase):
 
         target = Game.getObjectById(target_id)
 
-        if target is None or self.room.hostile:
+        if target is None or (self.room.hostile and target.owner.username != INVADER_USERNAME):
             self.memory.attack_target = None
             delete_target(target_id)
             return True
@@ -77,7 +77,6 @@ profiling.profile_whitelist(RoleDefender, ["run"])
 
 class MilitaryBase(RoleBase):
     def _find_nearest_junctions(self):
-
         room_xy = parse_room_to_xy(self.pos.roomName)
         if room_xy is None:  # we're in sim
             return []
